Refactor GUI components into its own package
