+++ src/main/java/duke/gui/MainWindow.java
@@ -1,7 +1,8 @@
-package duke;
+package duke.gui;
 
 import static duke.command.ByeCommand.BYE_MESSAGE;
 
+import duke.Duke;
 import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
